User ID none case, ++deny

+++ src/cmds/deny.py
@@ -27,6 +27,9 @@ async def perform_action(ctx: ApplicationContext, reply, ban_id):
             for row in cursor.fetchall():
                 user_id = row[0]
 
+            if user_id is None:
+                await reply(ctx, 'Cannot find record of ban request. Has this user already been unbanned?')
+                return
             await unban.perform_action(ctx, reply, user_id)
             query_str = 'DELETE FROM ban_record WHERE id = %s'
             cursor.execute(query_str, (ban_id,))
